project-zero/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
boto3>=1.35.0
Pillow>=10.0.0
httpx>=0.27.0